refactor: adjust package struct & build redux-ui
