--- packages/redux-ui/base/index.d.ts
@@ -1,13 +1,0 @@
-export class ReduxUiBase {
-  key: String;
-  reduxPath: String;
-}
-
-interface BaseAction extends Action {
-  meta?: Object;
-  key: String;
-}
-
-interface BaseOpts {
-  reduxPath?: String;
-}
