--- packages/redux-ui/index.d.ts
@@ -1,10 +1,0 @@
-import { Reducer } from "redux";
-
-interface ReduxUiReducers {
-  toggles: Reducer;
-  assigns: Reducer;
-  progresses: Reducer;
-}
-
-export const reduxUiReducers: ReduxUiReducers;
-
