+++ packages/redux-ui/package.json
@@ -10,27 +10,22 @@
   "publishConfig": {
     "access": "public"
   },
-  "template": "cra-redux",
+  "template": "redux-library",
   "files": [
-    "index.js",
-    "index.d.ts",
-    "base",
+    "dist",
     "bin",
-    "lib",
-    "typings",
-    "assign",
-    "progress",
-    "toggle"
+    "typings"
   ],
   "config-overrides-path": "../cli/config-overrides",
-  "main": "src/index.js",
-  "module": "index.js",
-  "typings": "index.d.ts",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
+  "typings": "typings/index.d.ts",
   "scripts": {
     "format": "prettier --trailing-comma es5 --write '**/*.js'",
     "lint": "sketch lint src",
     "release": "standard-version && git push --follow-tags origin master",
-    "test": "NODE_ENV=test sketch test --env=node"
+    "test": "NODE_ENV=test sketch test --env=node",
+    "build": "sketch build"
   },
   "commitlint": {
     "extends": [
@@ -50,7 +45,7 @@
       "git add"
     ]
   },
-  "dependencies": {
+  "peerDependencies": {
     "lodash": "^4.17.11",
     "redux-saga": "^1.0.2",
     "reselect": "^4.0.0"
