+++ packages/redux-ui/src/assign/action.js
@@ -1,4 +1,4 @@
-import { endsWith, startsWith } from "lodash";
+import { startsWith } from "lodash";
 export const PREFIX = "@ASSIGN";
 
 export const SET = "SET";
