+++ packages/redux-ui/src/index.js
@@ -7,3 +7,7 @@ export const reduxUiReducers = {
   assigns: assignReducer,
   progresses: progressReducer,
 };
+
+export * from "./toggle";
+export * from "./assign";
+export * from "./progress";
