+++ packages/redux-ui/src/progress/action.js
@@ -1,4 +1,4 @@
-import { endsWith, startsWith } from "lodash";
+import { startsWith } from "lodash";
 
 export const PREFIX = "@PROGRESS";
