+++ packages/redux-ui/src/progress/reducer.js
