+++ packages/redux-ui/src/toggle/action.js
@@ -1,4 +1,4 @@
-import { endsWith, startsWith } from "lodash";
+import { startsWith } from "lodash";
 export const PREFIX = "@TOGGLE";
 
 export const OPEN = "OPEN";
